Simplified redux setup
