+++ src/client/index.jsx
@@ -47,8 +47,6 @@ import ErrorFallback from './components/ErrorFallback'
 
 import { tasks } from './tasks'
 
-import { store, history, sagaMiddleware } from './middleware'
-
 function parseProps(domNode) {
   return 'props' in domNode.dataset ? JSON.parse(domNode.dataset.props) : {}
 }
@@ -96,12 +94,7 @@ function App() {
       FallbackComponent={ErrorFallback}
       onError={(error) => ReactSentry.captureException(error)}
     >
-      <Provider
-        sagaMiddleware={sagaMiddleware}
-        history={history}
-        store={store}
-        tasks={tasks}
-      >
+      <Provider tasks={tasks}>
         <Mount selector="#data-hub-header">
           {(props) => <DataHubHeaderWrapper {...props} />}
         </Mount>
